HSEARCH-4153 Refactor PojoIndexingPlan/PojoIndexer tests
So that we can more easily add more tests for loading
