+++ integrationtest/mapper/pojo-base/src/test/java/org/hibernate/search/integrationtest/mapper/pojo/work/operations/IndexedEntity.java
@@ -1,0 +1,31 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.integrationtest.mapper.pojo.work.operations;
+
+import org.hibernate.search.integrationtest.mapper.pojo.work.PojoIndexingPlanBaseIT;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.DocumentId;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
+
+@Indexed(index = PojoIndexingPlanBaseIT.IndexedEntity.INDEX)
+public final class IndexedEntity {
+
+	public static final String INDEX = "IndexedEntity";
+
+	public static IndexedEntity of(int id) {
+		IndexedEntity entity = new IndexedEntity();
+		entity.id = id;
+		entity.value = String.valueOf( id );
+		return entity;
+	}
+
+	@DocumentId
+	Integer id;
+
+	@GenericField
+	String value;
+}
